etl/requirements.txt
backoff==2.2.1
elastic-transport==8.13.1
elasticsearch==8.14.0
load-dotenv==0.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
urllib3==2.2.1